requirements.txt
telethon==1.38.0
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==10.0.0
openai==1.0.0
dateparser==1.1.8
tqdm==4.66.1
requests==2.31.0
google-generativeai==0.3.2
ratelimit==2.2.1